+++ server/src/__tests__/config.test.ts
@@ -13,6 +13,17 @@ describe('ConfigSchema', () => {
         "logLevel": "info",
         "shellcheckArguments": [],
         "shellcheckPath": "shellcheck",
+        "shfmt": {
+          "binaryNextLine": false,
+          "caseIndent": false,
+          "funcNextLine": false,
+          "ignoreEditorconfig": false,
+          "keepPadding": false,
+          "languageDialect": "auto",
+          "path": "shfmt",
+          "simplifyCode": false,
+          "spaceRedirects": false,
+        },
       }
     `)
   })
@@ -25,6 +36,17 @@ describe('ConfigSchema', () => {
         includeAllWorkspaceSymbols: true,
         shellcheckArguments: ' -e SC2001  -e SC2002 ',
         shellcheckPath: '',
+        shfmt: {
+          binaryNextLine: true,
+          caseIndent: true,
+          funcNextLine: true,
+          ignoreEditorconfig: true,
+          keepPadding: true,
+          languageDialect: 'posix',
+          path: 'myshfmt',
+          simplifyCode: true,
+          spaceRedirects: true,
+        },
       }),
     ).toMatchInlineSnapshot(`
       {
@@ -41,6 +63,17 @@ describe('ConfigSchema', () => {
           "SC2002",
         ],
         "shellcheckPath": "",
+        "shfmt": {
+          "binaryNextLine": true,
+          "caseIndent": true,
+          "funcNextLine": true,
+          "ignoreEditorconfig": true,
+          "keepPadding": true,
+          "languageDialect": "posix",
+          "path": "myshfmt",
+          "simplifyCode": true,
+          "spaceRedirects": true,
+        },
       }
     `)
   })
@@ -67,12 +100,24 @@ describe('getConfigFromEnvironmentVariables', () => {
         "logLevel": "info",
         "shellcheckArguments": [],
         "shellcheckPath": "shellcheck",
+        "shfmt": {
+          "binaryNextLine": false,
+          "caseIndent": false,
+          "funcNextLine": false,
+          "ignoreEditorconfig": false,
+          "keepPadding": false,
+          "languageDialect": "auto",
+          "path": "shfmt",
+          "simplifyCode": false,
+          "spaceRedirects": false,
+        },
       }
     `)
   })
   it('preserves an empty string', () => {
     process.env = {
       SHELLCHECK_PATH: '',
+      SHFMT_PATH: '',
       EXPLAINSHELL_ENDPOINT: '',
     }
     const { config } = getConfigFromEnvironmentVariables()
@@ -86,6 +131,17 @@ describe('getConfigFromEnvironmentVariables', () => {
         "logLevel": "info",
         "shellcheckArguments": [],
         "shellcheckPath": "",
+        "shfmt": {
+          "binaryNextLine": false,
+          "caseIndent": false,
+          "funcNextLine": false,
+          "ignoreEditorconfig": false,
+          "keepPadding": false,
+          "languageDialect": "auto",
+          "path": "",
+          "simplifyCode": false,
+          "spaceRedirects": false,
+        },
       }
     `)
   })
@@ -94,6 +150,8 @@ describe('getConfigFromEnvironmentVariables', () => {
     process.env = {
       SHELLCHECK_PATH: '/path/to/shellcheck',
       SHELLCHECK_ARGUMENTS: '-e SC2001',
+      SHFMT_PATH: '/path/to/shfmt',
+      SHFMT_CASE_INDENT: 'true',
       EXPLAINSHELL_ENDPOINT: 'localhost:8080',
       GLOB_PATTERN: '*.*',
       BACKGROUND_ANALYSIS_MAX_FILES: '1',
@@ -113,6 +171,17 @@ describe('getConfigFromEnvironmentVariables', () => {
           "SC2001",
         ],
         "shellcheckPath": "/path/to/shellcheck",
+        "shfmt": {
+          "binaryNextLine": false,
+          "caseIndent": true,
+          "funcNextLine": false,
+          "ignoreEditorconfig": false,
+          "keepPadding": false,
+          "languageDialect": "auto",
+          "path": "/path/to/shfmt",
+          "simplifyCode": false,
+          "spaceRedirects": false,
+        },
       }
     `)
   })