+++ server/src/config.ts
@@ -40,6 +40,37 @@ export const ConfigSchema = z.object({
 
   // Controls the executable used for ShellCheck linting information. An empty string will disable linting.
   shellcheckPath: z.string().trim().default('shellcheck'),
+
+  shfmt: z
+    .object({
+      // Controls the executable used for Shfmt formatting. An empty string will disable formatting
+      path: z.string().trim().default('shfmt'),
+
+      // Ignore shfmt config options in .editorconfig (always use language server config)
+      ignoreEditorconfig: z.boolean().default(false),
+
+      // Language dialect to use when parsing (bash/posix/mksh/bats).
+      languageDialect: z.enum(['auto', 'bash', 'posix', 'mksh', 'bats']).default('auto'),
+
+      // Allow boolean operators (like && and ||) to start a line.
+      binaryNextLine: z.boolean().default(false),
+
+      // Indent patterns in case statements.
+      caseIndent: z.boolean().default(false),
+
+      // Place function opening braces on a separate line.
+      funcNextLine: z.boolean().default(false),
+
+      // (Deprecated) Keep column alignment padding.
+      keepPadding: z.boolean().default(false),
+
+      // Simplify code before formatting.
+      simplifyCode: z.boolean().default(false),
+
+      // Follow redirection operators with a space.
+      spaceRedirects: z.boolean().default(false),
+    })
+    .default({}),
 })
 
 export type Config = z.infer<typeof ConfigSchema>
@@ -57,6 +88,17 @@ export function getConfigFromEnvironmentVariables(): {
     logLevel: process.env[LOG_LEVEL_ENV_VAR],
     shellcheckArguments: process.env.SHELLCHECK_ARGUMENTS,
     shellcheckPath: process.env.SHELLCHECK_PATH,
+    shfmt: {
+      path: process.env.SHFMT_PATH,
+      ignoreEditorconfig: toBoolean(process.env.SHFMT_IGNORE_EDITORCONFIG),
+      languageDialect: process.env.SHFMT_LANGUAGE_DIALECT,
+      binaryNextLine: toBoolean(process.env.SHFMT_BINARY_NEXT_LINE),
+      caseIndent: toBoolean(process.env.SHFMT_CASE_INDENT),
+      funcNextLine: toBoolean(process.env.SHFMT_FUNC_NEXT_LINE),
+      keepPadding: toBoolean(process.env.SHFMT_KEEP_PADDING),
+      simplifyCode: toBoolean(process.env.SHFMT_SIMPLIFY_CODE),
+      spaceRedirects: toBoolean(process.env.SHFMT_SPACE_REDIRECTS),
+    },
   }
 
   const environmentVariablesUsed = Object.entries(rawConfig)